Add logic to addimage

+++ src/Imi.Project.WPF/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using Imi.Project.WPF.Interfaces;
+﻿using Imi.Project.WPF.Events;
+using Imi.Project.WPF.Interfaces;
 using Imi.Project.WPF.Models.Birds;
 using Imi.Project.WPF.ViewModels;
 using Imi.Project.WPF.Views;
@@ -16,18 +17,15 @@ namespace Imi.Project.WPF
         private readonly ISpeciesApiService _speciesApiService;
         private readonly ICageApiService _cageApiService;
 
-
-        public delegate void RefreshList();
-        public event RefreshList RefreshListEvent;
-
-        public MainWindow(ISpeciesApiService speciesApiService, IBirdApiService birdApiService, ICageApiService cageApiService)
+        public MainWindow(ISpeciesApiService speciesApiService, 
+                          IBirdApiService birdApiService, 
+                          ICageApiService cageApiService)
         {
             InitializeComponent();
             _speciesApiService = speciesApiService;
             _birdApiService = birdApiService;
             _cageApiService = cageApiService;
             SetData();
-
         }
 
         private async void SetData()
@@ -50,37 +48,44 @@ namespace Imi.Project.WPF
         private void btnAddBird_Click(object sender, RoutedEventArgs e)
         {
             AddBird addBird = new AddBird(_birdApiService, _speciesApiService, _cageApiService);
-            RefreshListEvent += new RefreshList(RefreshBirdList); 
-            addBird.BirdAdded = RefreshListEvent;
+
+            addBird.BirdAdded += RefreshBirdList;
             addBird.Show();
         }
 
         private void btnEditBird_Click(object sender, RoutedEventArgs e)
         {
-            EditBird editBird = new EditBird(_birdApiService, _speciesApiService, _cageApiService, (BirdApiResponse)lstBirds.SelectedItem);
-            RefreshListEvent += new RefreshList(RefreshBirdList);
-            editBird.BirdEdited = RefreshListEvent;
+            EditBird editBird = new EditBird(_birdApiService, _speciesApiService, _cageApiService, 
+                                                            (BirdApiResponse)lstBirds.SelectedItem);
+
+            editBird.BirdEdited += RefreshBirdList;
             editBird.Show();
         }
 
         private async void btnDeleteBird_Click(object sender, RoutedEventArgs e)
         {
-            var birdToDelete = (BirdApiResponse)lstBirds.SelectedItem;
-            var result = await _birdApiService.DeleteBird(birdToDelete.Id);
-            if (!ReferenceEquals(result, null)) MessageBox.Show($"Something went wrong\n{result}", null,
-                                                         MessageBoxButton.OK, MessageBoxImage.Exclamation);
-            else
+            var action = MessageBox.Show($"Are you sure you want to delete this bird?", "Hold!",
+                                                         MessageBoxButton.YesNoCancel, MessageBoxImage.Warning);
+            if (action == MessageBoxResult.Yes)
             {
-                MessageBox.Show("Bird was succesfully deleted!", "Success", MessageBoxButton.OKCancel, MessageBoxImage.Information);
-                SetData();
-                lstBirds.SelectedIndex = -1;
-                stkDetails.Visibility = Visibility.Hidden;
+                var birdToDelete = (BirdApiResponse)lstBirds.SelectedItem;
+                var result = await _birdApiService.DeleteBird(birdToDelete.Id);
+                if (!ReferenceEquals(result, null)) MessageBox.Show($"Something went wrong\n{result}", null,
+                                             MessageBoxButton.OK, MessageBoxImage.Exclamation);
+                else
+                {
+                    MessageBox.Show("Bird was succesfully deleted!", "Success", 
+                                         MessageBoxButton.OKCancel, MessageBoxImage.Information);
+                    SetData();
+                    lstBirds.SelectedIndex = -1;
+                    stkDetails.Visibility = Visibility.Hidden;
+                }
             }
-
         }
-        private void RefreshBirdList()
+        private void RefreshBirdList(object sender, BirdAddedOrEditedArgs e)
         {
             SetData();
+            if (e.Bird != null) lstBirds.SelectedItem = e.Bird;
         }
     }
 }
